ENH: 위경도 data type 변경

+++ src/main/java/topia/duck/hack/controller/dto/request/SubTopicCreateDto.java
@@ -26,10 +26,10 @@ public class SubTopicCreateDto {
     private String description;
 
     @JsonProperty("latitude")
-    private double latitude;
+    private long latitude;
 
     @JsonProperty("longitude")
-    private double longitude;
+    private long longitude;
 
     @JsonProperty("address")
     private String address;
